+++ react/rangeslider/src/RangeSlider.tsx
@@ -1,0 +1,6 @@
+import { CoreSlider } from '../../slider/src/core/CoreSlider.tsx';
+import type { RangeSliderProps } from '../../slider/src/props.ts';
+
+export function RangeSlider(props: RangeSliderProps) {
+  return CoreSlider(props);
+}
