+++ react/slider/src/core/Tooltip.tsx
@@ -1,5 +1,5 @@
 import type { RefObject } from 'react';
-import { thumbWidth, type Ref } from './Slider.tsx';
+import { thumbWidth, type Ref } from './CoreSlider.tsx';
 
 // Tooltip component that shows a given value above the slider thumb.
 export const Tooltip = ({
