+++ react/slider/src/core/math.ts
