+++ react/slider/src/core/style.css
