--- react/slider/src/index.ts
@@ -1,2 +1,0 @@
-export { Slider } from './Slider.tsx';
-export type { Props } from './props.ts';
